fix: ProductImageService.php relocated to Utils and fixing references
